Reduce RepositoryArchive to a model and move to models/
